from __future__ import division
- proper use of true and floor division operators
- remove unnecessary int/float conversions

+++ demos/demos/apps/ea/tasks/pdf.py
@@ -1,5 +1,7 @@
 # File: pdf.py
 
+from __future__ import division
+
 import math
 
 from os import path
@@ -165,12 +167,12 @@ class BallotBuilder:
     page_width -= w_margin * 2
     page_height -= h_margin * 2
     
-    img_size = int(page_width // 4.5)
+    img_size = page_width // 4.5
     font_size_tag = int(img_size)
     
-    url_indent = int(page_width // 20)
-    table_top_gap = int(page_width // 15)
-    table_opt_gap = int(page_width // 50)
+    url_indent = page_width // 20
+    table_top_gap = page_width // 15
+    table_opt_gap = page_width // 50
     
     long_vc_split = 4
     long_vc_hyphens = int(math.ceil(config.VOTECODE_LEN / long_vc_split)) - 1
@@ -624,8 +626,8 @@ class BallotBuilder:
                         opt_rows = min(2*(avail_rows-1), data_len-row)
                         incl_ftr = (row + opt_rows == data_len)
                         
-                        frst = int(row // 2)
-                        last = int(2*frst+opt_rows - (2*frst+opt_rows)//2)
+                        frst = row // 2
+                        last = 2*frst+opt_rows - (2*frst+opt_rows)//2
                         
                         t = int(math.ceil(data_len/2))
